adding getRange to helpers
